requirements.txt
aiohttp==3.8.1
beautifulsoup4==4.10.0
discord.py==1.7.3
feedparser==6.0.8
mysql-connector==2.2.9
redis==4.1.4
python-dateutil
